requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
google-generativeai
python-multipart
python-dotenv
pillow
